Add capybara end-to-end tests
This is a simple test of using capybara with chrome for end to end
tests.

The actual tests don't do much - fill out a form and log into the admin
interface, both hard coded to staging.

A dockerfile is included to get chrome setup easily. It can be run
locally by downloaded
[chromedriver](https://chromedriver.chromium.org/downloads) directly and
placing it on the current path.

The setup and config is a bit messy - things could be split out and
formatted. It doesn't seem worth it unless we decide to use it though.

Instructions for use can be found in `capybara/README.md`