Deploy RDS Postgres Cluster to Development
Creates a module to manage an RDS Serverless V1 cluster with postgres 11
engine. Serverless V1 is compatible with AWS Data API and scales to
"off" when not in use (depending on configuration).

Adds a deployment to deploy the RDS module to the development
environment.

The `prepare.sql` is a basic means to setup the database and app users
and roles as a one-off to be run via the AWS Data API after the initial
creation.