Update and Correct Access to Deploy ECR
The ECS services running in development, staging and eventually
production pull images from the deploy account's ECR. This access is
controlled via an ECR policy which needs to grant the ECS task execution
roles for each of the services in each of the environments access.

Corrects the formatting to correctly apply a list of principals.