Shorten Target Group Deregistration Delay To 60 Seconds
The target group deregistration delay governs how long the target group
waits for connections to drain before removing an in-active ECS task from the
group. The default is 5 minutes (300 seconds) which makes our
deployments unnecessarily long. This should be fine-tuned later but 60
seconds provides sufficient time for in-flight requests to complete.