Restructure the Pipelines
Bring the deployment and subsequent running of the end-to-end tests into
the same CodePipeline Stage per environment being deployed to.

Update the CodeBuild end-to-end test projects to use different SSM
parameters for each environment. This is necessary since Notify API keys
are different and in future the Sigon Auth credentials will be
different.