Add Time Stamp to End-to-End Form Name
Adding the time stamp differentiates the form created for each run and
avoids issues when an end-to-end test fails before deleting the form it
created which then has knock on impact to following runs since there is
more than one match for the form name.

We might want to investigate making the end-to-end tests more resilient
and deleting the form when an error is encountered but this approach
also guards against the process existing unexpectedly.