Update for rails-api
The API project has moved from Grape to Rails and has a new Dockerfile
based on the other Rails projects.

This commit changes the docker compose environment to be compatible with
the new API.

It should only be used when the Rails API code is merged in to the main
branch of the API project.