Update Forms-Api Environment Variables
As part of migrating forms-api from Grape to Rails the way environments
variables are defined is changing. This updates the names accordingly
and adds in the Sentry related environment variables which will be
picked up once the rails version is deployed.

To distinguish between existing PaaS and new AWS environments the sentry
environment is prefixed with `aws`.