Add Env Vars to Forms-API
Following the migration to Rails framework forms-api now needs the
`SECRET_KEY_BASE` env var set, even though we don't really make use of
this Rails feature. We'll investigate further in the future to remove it
if possible.

The same approach of setting it to 'something' was taken for forms-admin
on aws environments.