Add Deployment of Apps To Production
Deploys the three apps to the production environment after a successful
deployment to the staging environment. The end-to-end tests are run
after deployment and in future we will alert if they fail.

Deploys the deployer-role module to production to create a role that can
be assumed by the CodeBuild projects to apply the ECS Terraform to the
production account.