Add AWS Data API Helper Script
AWS RDS Serverless V1 clusters can be queried via the AWS Data API.
Queries require the RDS Cluster ARN and SecretsManager secret ARN
holding the necessary Postgres username and password. This script looks
up those values based on the desired database and then executes the
provided query.

The script require confirmation that the user acknowledges that write
statements can also be run. It requires a second confirmation if the
statement would be run against production.

Permissions to perform the actions included in this script are managed
via IAM and must be run in an authenticated shell using gds-cli or
aws-vault. This script does not permit a user from doing anything they
do not already have permission to do.