Deploy Applications To User Research
After deploying to development environment update the application in the
User Research environment.

This will keep the environment current. The automatic deployment can be
disabled if necessary. Smoke tests are not included because we need to
implement basic auth and update the end-to-end tests first (if we feel
running the end-to-end post deployment into user research is necessary
in future).