Script To Build and Upload Application Docker Image
Occassionally we will need to deploy local source code for an
application into the development or user-research environment. To
achieve this we must:
1. build the local source code into a Docker image and push it into the
   ECR registry within our deploy account.
2. Apply the application Terraform to update the ECS service in the
   choosen environment, either development or user-research, to use the
   new image.

This script helps to accomplish the first task which is multistep
process and benefits from being automated. The second step is a simple
`terraform apply` operation. For now the frequency we'll be pushing
local source code to either environment is low but we can revisit this
need at a later date and consider CodePipelines to fully automate
deploying from development branches.