Add CloudFront Distribution
CloudFront is an AWS provided CDN. It provides benefits for dynamic
content regarding improved performance, availability and security.

This introduces a CloudFront distribution where the existing Application
Load Balancer is the origin.

The ACM Certificate for the CloudFront distribution must be in us-east-1
hence the addition of a new aliased AWS provider for that region.

The ALB is still reachable directly on its own Domain name e.g. ending
.eu-west-2.elb.amazonaws.com and we might want to restrict ALB from
forwarding requests unless they come via the CF CDN. Note that the SSL
certificate on the ALB is not valid for the AWS provided domain name
and so cannot be accidentally queried directly.