Turn Off Detect Changes on Development Branch Pipelines
Until set otherwise by a developer the dev branch pipelines just look at
main branch. There is no need for these pipelines to automatically
deploy on merges to main because the main-branch pipelines do that.

When a developer wants to deploy their code to development or
user-research environments they will:
1. Update the pipelines to point at their development branch
2. Optionally enable detect changes on the source code repo to
   automatically deploy whenever they commit, or leave it disabled and
   manually click "Release Change" when ready to deploy.