Add Basic WAF with Rate Limit ACL
A simple rate limit rule based on the number or requests from an IP in
the previous 5 minutes. The time period is not configurable but the
limit is, it has been set to 1,000 by default. The rule returns a 429
'Too many requests' when it kicks in.

The ACL will log details about blocked requests to a CloudWatch log
group which can be integrated with Splunk at a later date.