Helper Script to Update Basic Auth Creds in User Research
We need to periodically update the basic auth username and password in
the user-research environment. These credentials are held in AWS SSM
Parameter Store and retrieved by AWS ECS each time it launches a new
task for forms-admin.

This script makes updating the username and password simpler by updating
Parameter Store values and then redeploying forms-admin in the
user-research environment.