Create and Use Custom CodeBuild Environment For End-to-end Tests
Creates a Docker image with ruby installed to a version which we have
full control over. It also installed Chrome and Chrome Driver which are
necessary for the end-to-end tests. This has two main benefits over the
AWS provided CodeBuild image:
- We can set the Ruby version as we need and it won't change without us
  doing so (this previously broke the pipeline when the AWS CodeBuild
  image installed 3.1.3 instead of 3.1.2.
- The tests run faster because they do not install Chrome and Chrome
  Driver every time.

The custom environment runs as the `codebuild` user to avoid using
`root` and following Docker best practice.

A new CodePipeline named 'utilities' manages the building of the new
custom image and pushing into ECR. We may want to run this on a regular
schedule. This pipeline will be extended to build custom images for our
other projects in future.

The CodeBuild projects that run the end-to-end tests are updated to use
this new custom image and initial testing shows they run about 2 minutes
faster each time.

The CodeBuild project that builds the docker images is slightly modified to
permit passing a static image tag rather than creating one based on a
GIT_SHA and time stamp.

The ECR repo for the end-to-end test image permits mutable tags to
reduce the maintenance overhead of bumping each pipeline to pick up
updates to the image when we re-run the utilities pipeline.