Add PenTester Role
This role meets Cyber's requirements for Pen Testing. It has also been
granted to Dan and Alice to help debug any access permissions during
testing. Felix is waiting for their gds-users IAM user to be created but
is scheduled to be working on the Pen Testing (the role cannot be
granted until their IAM user is set up).