Set WAF Rate Limit Rule to Count Mode
"COUNT" mode permits the request but logs that it would have been blocked
if the rule was set to "BLOCK". This helps us to trial the rule without
actually blocking users.

When we are satisfied that the rule is correctly not blocking legimate
access and we have considered what the blocked users see we can set the
mode to "BLOCK".