Update content of delete confirmation page
Our end to end tests are failing on the changes in
alphagov/forms-admin#358, because the page h1 has changed. This commit
updates the test to look for either the new content or the old content,
so it should pass before and after forms-admin has been fully deployed.

After forms-admin has been fully deployed we can change the test to look
for just the new content.