Replace inline policy with managed policy
A recent pentest highlighted the use of this inline policy which is "poor
practice and is strongly discouraged"

For this to work we need to give the deployer role permissions to manage
policies (it was previously managing only _role_ policies). This is the role
that ECS assumes when starting a new app, for example to access the relevant
secrets from the parameter store.

We've also renamed resources that referred to parameter_store_exec_policy to
allow for more flexibility when adding other policies in the future. Instead
we've referred to them as "additional".

Co-authored-by: alice-carr <alice.carr@digital.cabinet-office.gov.uk>