Add Alerts Module
Creates an SNS topic with a PagerDuty subscriber. Notifications to this
topic are sent to PagerDuty. The PagerDury url endpoint is stored as a
secure string in Parameter Store and can be found in PagerDuty at
Services > GOV.UK Forms > Integrations > CloudWatch.

This initial commit creates a CloudWatch alarm to alert when the number
of healthy instances of our apps falls below the specified threshold.

The alert is applied to Staging and Production.