Use an external-id when the deployer assumes a role in another environment
The `deployer` will have to provide an external id when assuming a role e.g. in
a `dev` environment. The id will be the same as the name of the environment.

We're doing this because a security assessment recommended it. It is usually
relevant when you can have a confused deputy
(https://docs.aws.amazon.com/IAM/latest/UserGuide/confused-deputy.html) but we
don't think it is the case for us. However, we're implementing it since it is
trivial.