Ruby Script to Compare Forms in PaaS vs AWS
As part of our migration from Paas to AWS, we have created a script that Queries the forms-api in AWS and PaaS and checks if the form exists in
either PaaS, AWS or both. This will be used during the migration, to ensure that no forms are lost or changed during the switch over.

This script first checks all the base forms, found using the /forms endpoint, and then checks all live forms and draft forms using the live and draft
endpoints. The script will then print out if each form is identical in both AWS and PasaS, if a form is only present in PaaS and not AWS, or if there is a
discrepency in the content of the forms between the two platforms.

This can be run against development, staging and prod envirnments.

Co-authored-by: Catalina Garcia <catalina.garcia@digital.cabinet-office.gov.uk>
Co-authored-by: Dan Worth <dan.worth@digital.cabinet-office.gov.uk>
Co-authored-by: Radha Kotyankar <radha.kotyankar@digital.cabinet-office.gov.uk>