Fix Deletion Assertions on End-To-End Tests
If the environment doesn't have any existing forms before the end-to-end
tests run then after deleting the test form there will still be no forms
and hence the `govuk-table` element is not shown after deleting the test
form.

This changes the assertions so that it checks that the deletion message
is correctly presented in all cases. If there are then forms to display
it confirms that the deleted form does not appear in the list.