Script to Aid Updating Ruby and Alpine Versions
We need to keep Ruby and Alpine regularly updated in our apps
forms-runner, forms-admin and forms-api. This scripts automates the
process of updating the necessary files to make it simpler and faster.

Some consideration has been given to capturing errors but the
responsibility remains with carefully reviewing the generated commits.

Can be extended to manage version of Node where necessary.