Add auth_provider variable for forms-admin module
We want to be able to select from different authentication strategies
for the forms-admin app, including basic auth, GOV.UK Signon, and in
future auth0 or others.

This commit replaces the `enable_basic_auth` variable with a more
generic `auth_provider` variable, and also updates the environment
variables used in line with the changes to the environment variables
used in forms-admin in PR alphagov/forms-admin#461.

We also make change the secrets list so that the GDS_SSO secrets
are not provided to the app environment when they are not needed.