Add parity checker and refactor
We've added step to check if the databases match to be part of the
clear-and-dump script. The environment will now be inferred from the AWS
account that the user is logged into - hopefully this will help avoid human
error.

Users will now have to pass in the keys for the forms-api in PaaS and AWS