Script to Check DNS Resolution During CutOver
This script will check the domains used in production or staging and
print out whether they resolve to PaaS or AWS environments. To be used
during cut-over between PaaS and AWS to simplify the checkout process.

Note:
- PaaS development environment does not have custom domains set
up so is not considered in this script
- There is no user research environment on PaaS and is also not
  considered int his script.
- The product pages only have a custom domain of
  www.forms.service.gov.uk in the production environment. We are not
  migrating them at this point in time so the script just checks that
  they're still reachable by the custom domain which shouldn't change.

DNS Check Script

Use the term 'permanent_domains' instead of 'real_domains'
Correct the domain for forms-api in staging