Fix or ignore outstanding lintint issues
Ran `bundle exec rubocop --autocorrect-all --disable-uncorrectable` to
fix or ignore linting issues that 'unsafe' autocorrects [[1]].

Most of these are around how we write tests. We could probably rewrite
the tests to reduce/avoid the use of `described_class` and
`allow_any_instance_of`, but for this changeset I thought it would be
better to just keep things simple.

[1]: https://docs.rubocop.org/rubocop/1.52/usage/autocorrect.html#example-of-unsafe-cop