Split out data API functionality from command into separate class
I want to be able to write a script that uses the AWS RDS data API to
communicate with deployed databases; to make this easier it would be
nice if I could reuse some of the code in the `data_api` command. This
commit splits the code that actually sets up the connection parameters
and does the communicating into a new helper class, `DataApiConnection`.