Add Staging Hosted Zone
Add Route53 hosted zone to the staging account for the
staging.forms.service.gov.uk domain. This is the permenant domain that
will be used after the migration from PaaS. For now it includes records
which resolve admin and submit subdomains to the PaaS CloudFront
distributions.

Add an NS record in the production hosted zone to delegate
staging.forms.service.gov.uk to the newly created Hosted Zone above.

Rename some resources to make it clearer.