Manage CloudFront Records in DNS Deployment
Move the creation of the CNAME records for the CloudFront distributions
in the dns deployment rather than create them when the CloudFront
distributions are created. To enable the migration we need finer control
of resolving domains to either the PaaS or AWS CloudFront distributions.

This also nicely separates the concerns between creating the CloudFront
distribution and resolving a domain to it.

The net effect once all Terraform has been applied is zero, its just
moving around where we manage the CNAME relating to the CloudFront
distribution.