Resolve Domains to ALB
A limitation within AWS prevents two CloudFront distributions having
overlapping subject alternative names. To make cut over simpler we will
temporarily remove the CloudFront distributions from our AWS
environments and resolve the DNS to the ALB directly.

The main purpose of the CloudFront distribution at this point was for a
WAF rule to rate limit traffic from a single IP. This rule is still in
COUNT mode and not intended to be set to BLOCK prior to the cutover.

We will reintroduce the CloudFront distributions and enable the rate
limiting rule after the migration when there is no longer an issue with
conflicting CloudFront distributions.