Update DNS Check Script For ALBS
We are temporarily removing CloudFront and resolving the domains to the
ALB instead. This updates the script to check if the domains resolve to
the ALB rather than CloudFront distribution.

I have left the function for finding the IP addresses of the CloudFront
distribution in since we'll be reverting to that in the near future.