Add data migration for organisation ID
We want to use the organisation ID instead of slug to associate
organisations with forms in the forms-api database; however, to do this
we need to get data from the forms-admin database, where the
organisations are stored.

This commit adds a script to do this. It should only be run once per
environment, but I thought it would be better to commit it so that it
could be put through the normal pull request review process.