Support Role Can Manage Main Branch Deployments
The support role can now disable and enable stage transitions on the
main branch pipelines to any environment (previously it was limited to
only development and user-research). This is necessary so that engineers
can pause deployments to any of the environments for example during an
incident stopping deployments to production.