Allow support role to perform ecs:StopTask
There may be occasions where a dev assuming the support role needs to
stop a task; for instance, if they have started a command with
`run-task` and then realised they have made a mistake and want to stop
it.

Unfortunately it looks like we can't restrict the task resource beyond
cluster (which is too broad) and task ID (which we can't know ahead of
time), so this policy just allows the support role to stop any task. But
with services a stopped task will be resurrected, so this shouldn't be
too bad if assumed by a bad actor.