Allow ACM Creation in Different Regions
CloudFront requires its certificate to be in us-east-1 whilst the rest
of our infrastructure and ALB certificate is in eu-west-2. This change
makes use of Terraform provider functionality by passing multiple aws
providers to the acm-cert-with-dns-validation so that calling modules
can control where the certificate is created.

This also solves a dependency issue and calling modules now correctly
wait until the certificate is validated by making the output depend on
the aws_acm_certificate_validation resource.