Add run_task command
Adds a command to forms-cli to make running a task in ECS easier; given
an app to run the command against it takes care of picking the right
VPC, subnets, and security groups.

It will also start printing the logs to the console, for easy
monitoring.

For now we intentionally restrict it to running only rails commands; the
intended use case is running a rake script that is defined in the apps
`lib/tasks` folder.