Fix run_task policy to allow running latest active task definition
For some reason using a wildcard for the revision in the task definition
arn doesn't allow leaving off a revision when calling RunTask, which is
the easier thing to do when calling the API from a script.

We need both, because when running a task via the AWS web console, the
revision will always be included, even when running the latest active
revision.