Add previous_auth_provider
When changing auth providers, between gds_sso and auth0 for example,
users who are logged in with the previous auth provider are unable to
logout after the change.

This happens because the environment variables for the auth_provider are
no longer available and so the application is can no longer build links
for logout correctly.

This commit adds a new variable, previous_auth_provider, which can be
set to preserve environment variables. For example to migrate to auth0
from gds_sso in development:

infra/deployments/development/forms-admin/main.tf

```terraform
  auth_provider                 = "auth0"
  previous_auth_provider        = "gds_sso"
```