S3 Product Pages Bucket
AWS Route53 does not support creating CNAME records for the apex of the
hosted zone (we need this so that we can direct forms.service.gov.uk to
our product pages on www.forms.service.gov.uk).

A workaround is to use an S3 bucket with the name of the apex and set it
to redirect all requests to the necessary sub-domain. This was originally
set up in production via the console so this commit will:
- Create the necessary S3 bucket in each pre-production environment so
  our environments all work the same.
- Manage the existing production S3 bucket in Terraform after we've
  imported it.