Add env var to the product pages to serve assets
The terraform for the product pages is not setting the environment
variable RAILS_SERVE_STATIC_FILES. This means that the built assets,
such as CSS, are not being served by the docker image. This results in
the product pages being unstyled.

This commit adds the environment variable, which we use in the Admin and
runner too.

There may be more sophisticated ways to serve assets in production in
the future - e.g. CDN or dedicated file server. We made the decision to
use the Rails serving assets because it makes the docker images self
contained and is simple.