Add libyaml dependency to e2e test docker image
https://github.com/alphagov/forms-e2e-tests/pull/5 added "debug" gem which
has a dependency called psych. psych need to have libyaml install to be able to
compile. This isn't an issue in our main repos forms-admin, forms-runner or forms-api
and this maybe due to the fact that those docker images are based off alpine images
and not aws ones like this Dockerfile.