Add Auth0 tenant
Adds a tenant to the auth0 module. An Auth0 tenant corresponds to a
deployment, it is the root of the deployment, and holds some general
configuration [[1]].

We don't actually need to have the tenant managed in Terraform, the
Auth0 Management API only ever applies to one tenant, based on the API
token used. However I thought it would be useful because then we can
have the friendly_name as a "canary", say if you want to test some
changes in the dev tenant first but are using the production tenant API
keys, then you would see in the Terraform plan that the tenant friendly
name would be changed and see that you are running Terraform against the
wrong tentant.

Note: tenants cannot be created with the Auth0 Management API yet [[2]],
so for the development deployment we've also created a tenant by hand
and run `terraform import` to add it to the state file.

[1]: https://auth0.com/docs/get-started/auth0-overview/create-tenants
[2]: https://registry.terraform.io/providers/auth0/auth0/latest/docs/resources/tenant