Customize Auth0 prompts
This commit configures the user experience of signing in to Auth0 the
way we want it. Our designers @hazalarpaliki and @hannahkc configured
the universal login to look like the GOV.UK Design System and follow the
GOV.UK content style guide in the Auth0 web dashboard; this commit
imports those changes into Terraform. It's not a perfect facsimile of
GOV.UK Frontend, but it's good enough for now, and we can iterate on it.

The configuration was imported from our existing `dev-passwordless` instance
using the Auth0 Management API Explorer [[1]].

For content, we have placed it in individual JSON files to make the
Terraform module file smaller. The auth0_prompt_custom_text resource
wants the configuration as a JSON string anyway [[2]].

We took care to replace the Unicode WORD JOINER character in GOV.UK with
a JSON escape sequence to make it easier to see and edit, using:

```
sed "s/\xe2\x81\xa0/\\\\u2060/g"
```

[1]: https://auth0.com/docs/api/management/v2/prompts/get-custom-text-by-language
[2]: https://registry.terraform.io/providers/auth0/auth0/latest/docs/resources/prompt_custom_text