Configure Terraform to copy Auth0 app secrets to AWS for us
When running the forms-admin app with the auth provider set to Auth0, it
needs the Auth0 secrets in the AWS parameter store at deployment time.
We could copy and paste these ourselves, but it's better to get
Terraform to do it for us.

Note that this does mean the secrets are stored in the statefile,
however as we are using the s3 backend they are stored encrypted on the
server [[1]] and only kept in memory when Terraform is running [[2]], so
this isn't too much of an issue.

[1]: https://docs.aws.amazon.com/AmazonS3/latest/userguide/UsingServerSideEncryption.html
[2]: https://developer.hashicorp.com/terraform/language/state/sensitive-data