Alias Records For Apex to The ALB
Route the apex of each environment's domain to the ALB so that we can
serve the product pages from the apex.

Note that the checkov rule has been skipped because it does not appear
applicable to A records using the alias block (which conflicts with the
`records` attribute I believe the rule highlights is missing).