Permit Support Role to Manage Parameter Store
Creating, changing and deleting parameter store values is a common
support task and shouldn't require admin role to perform.

The permissions are allowed on any parameter since its also necessary to
create values outside of the usual application name spaces (e.g.
/forms-api).

The password for the RDS database is excluded from this permission since
it should not change under normal support activities and if it was
accidentally deleted or modified it would be problematic.