Add e2e tests Auth0 env vars from parameter store
Switching to auth0 from signon means the end to end tests have changed
to include a new login method.

When running the end-to-end tests in codebuild we need to access the
parameter store to get the new environment variables.

This commit adds the three new variables which are fetched for each
environment.

The values have already been put in parameter store.