Use template file for passwordless email content
We want the from address for the passwordless email to depend on the
deployment. To do this we use the Terraform `templatefile` function.

Note that this means there are now two template syntaxes in play in the
passwordless email content; the Terraform template syntax using `${}`,
and the Liquid template syntax with `{{}}` used by Auth0. Fortunately
the two templates have different ways of marking interpolation, but care
is needed to make sure the right one is being used. The Terraform
interpolation happens when `terraform apply` is run, whereas the Liquid
template interpolation happens after, in Auth0.