Refactor engineer access roles to make access heirarchy implicit
This commit moves the logic that 'admin users have support user
permissions' into the module, rather than making it a concern of the
module caller.

Arguably this is less explicit, however the way the roles are structured
is arguably an implementation detail. This change also has a number of
benefits; as well as removing repetition from user lists, it reduces the
chance of error (it doesn't make sense for a user to have support access
and not readonly access), and it makes sharing the list of users
simpler.