Permit Support To Manage Dashboards
We create and edit CloudWatch dashboards in the console and then export
the JSON into a Terraform module to save it in case we need to reset or
recreate it. This is  because it much simpler to edit and create in the
console than by trying to edit JSON representation directly. We should
be able to do this without needing to assume the admin role which has
higher auditing and alerting.