Correct SQS Policy and Enable SSE
The SQS policy needs to specify SNS as a service principal. Also enable
SSE with customer managed KMS key on SQS and SNS. A customer managed Key
is necessary since SES needs permission to use the key and this cannot
be granted on the default key that SNS uses.

The KMS key policy must permit the role applying the change to manage
the key otherwise it will fail to apply; this is to prevent accidentally
locking ourselves out of managing the key. Therefor each admin has
admininstractive permissions. The support role and the admin role can
use the key to decrypt messages so that they can see what is on the SQS
queue and deal with them. Unfortunately this means spreading our
engineer access our from the engineer-access deployment and into this
SES deployment.