wip! Use app setup scripts in command
This lets us simplify the default command for each container, and the
init.sql script for setting up the PostgreSQL database, as for each app
part of the `bin/setup` script is creating the required database if it
doesn't already exist.

Note that we remove from `init.sql` the lines about granting privileges
to the database, but they are unneeded as the user we give the user we
create for each app the ability to create databases, and when a user
creates a database they will automatically be the owner and have all
privileges already.