Fix configuration for inter-app communication
Currently forms-admin is not able to talk to forms-api or forms-runner
and forms-runner is not able to talk to forms-api.

This is because the environment variables used to configure the
forms-api and forms-runner URLs and keys have been changed in the
forms-admin and forms-runner apps since the configuration in this
Compose file was written [[1], [2], [3]].

This commit updates the environment for each container in docker-compose
to have the right variables.

[1]: https://github.com/alphagov/forms-admin/pull/334
[2]: https://github.com/alphagov/forms-runner/pull/145
[3]: https://github.com/alphagov/forms-runner/pull/247