Use build target from app Dockerfiles for Docker Compose images
We need to be able to install development gems in our Docker Compose
containers; otherwise we aren't able to run `rails db:seed` or `vite` in
our Docker Compose stack.

To do this we need some development dependencies installed in the
system. In our production images those headers and libraries are removed
after they are used; fortunately however our Dockerfiles use layers to
do this, and Docker Compose lets us pick which layer to use for our
images.

This commit tells `docker-compose build` to use the `build` layer for
each of our app images. This layer has the all the dependencies needed
to build and install the development gems and node modules.

This also has the advantage of saving a bit of time when running
`docker-compose build`, as there is less to do for each app image. So we
make this change for all apps, not just the ones where it is needed to
fix things right now. There is a small downside which is that we now
need to specify a command for each app service, but that is easily
added.

Note that the start command we use for forms-api is what used to be in
the Dockerfile before alphagov/forms-api#359, which moved the
responsibility of migrating the database from forms-api to the
deployment pipeline, because otherwise an extra command is needed by the
user to prepare the system for use.