Parametrise DNS configuration for SES module
Move the DNS records needed for the SES module from the deployment
`main.tf` file into the module itself, so that the module consumer
doesn't need to think about them.

No changes should need to be applied as a result of this change, as the
state has been moved using the Terraform `state rm` and `import`
commands.