Ensure that custom log events are sent to Splunk
The filter pattern used to exclude logs from the `ping` endpoint
inadvertantly also filtered out our custom log events. This commit fixes
that.

This happened because the filter syntax clause `{ $.action != "ping" }`
only matches events with a field `action`. Any events without that field
are filtered out; this includes our custom log events from the
forms-runner app [[1]]. To fix this we need to include events without
the field in our filter; the way to do this is buried deep in the
documentation [[2]], but fortunately it is possible.

[1]: https://github.com/alphagov/forms-runner/blob/1e23971cef072a21e976085a38e4b3f823ddee7a/app/lib/event_logger.rb
[2]: https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/FilterAndPatternSyntax.html