Improve feedback when shellcheck-ing scripts in GitHub Actions
Previously the snippets would be fed into Shellcheck on stdin, and errors would
show just the first line of the input. This isn't very helpful when trying to
find and fix the errors it's showing.

Insted we should show the YQ path to the snippet so that it can be more easily
found. We do this by extracting each snippet to a file whose name contains the
workflow file and the YQ path, so that when Shellcheck errors it outputs enough
information.