Update to amazonlinux:2.0.20231012.1
Previously we pinned to a hash. This was a good thing in the sense that we
could always know exactly what we were getting, but bad in that it made it hard
to know if the version we have today is older or new than some other version.

We switch to using Amazon's dated-based image tags to make it easier for us to
compare versions in the future. There is the chance that the tag will change to
point to a different hash, but we already trust Amazon with so much more that
it's a safe enough bet to trust them to not do/allow anything malicious to
happen with the container image tags.