Ensure we always update images everywhere
We use the same image in multiple plces, and it's very possible for us to use
different tags in different places. This test tries to discover every container
image we're using in Terraform, and subsequently tests that each image uses the
same tag. For container images that have their tag specified as a variable
('image:${tag}') we're comparing the lexical value of the tag ('${tag}')