Rename parameter group and create new one
We want to update Redis to engine version 7.0, so we need a parameter group with
the corresponding parameter group family (7). However, we cannot just change the
family for the existing parameter group as terraform tries to create a new one
with the same name.

Instead we can create two parameter groups, one for each version, and refer to
them by the version tag