Enable a shell script to be run prior to the deployment of an ECS service
`forms-api` needs to perform database migrations as part of its deployment
process. Previously, it was done in the startup command of the container,
however this caused contention for the database lock when 2 or more containers
started simultaneously. When this happened, all but 1 of the containers would
fail and exit, and ECS would restart the failures until they all started
successfully. This resulted in an ever increasing number of container restarts,
as the number of containers grew.

What we actually want is for database migrations to happen a single time, no
matter how many container instances there are. That means the migration code
needs to run before all of the containers are scheduled. To implement that we
have a few considerations:

1. The code should run at the point in the process that most people would expect it
2. The change should introduce as little complexity as we can get away with
3. The solution should be one we can re-use for other similar scenarios
4. The solution should avoid adding new tools unless absolutely necessary

We settled on allowing a shell script to run during `terraform apply`, after a
new ECS task definition with the new container image (containing the database
migrations) is created, and before the new containers are started up.

Running arbitrary code is not something Terraform handles well, so we've relied
on a combination of the "null_resource" resource type and the "local-exec"
provisioner to inform Terraform of a new resource that represents the running
of the script.

The script itself uses the AWS ECS RunTask API action to run a container, based
on the ECS task definition, with a different start command, then waits for it
to finish using the `aws ecs wait tsks-stopped` CLI utility. The RunTask API is
different from other parts of the AWS ECS API in that it does not attempt to
restart a contaienr when it exits, which is exactly the behaviour we're looking
for.