Set the version of Terraform being used in CodePipeline during upgrades
Our CodePipeline pipeline configurations take in the version of Terraform that
they'll need to use. We want it to be overridable for the times we need it, so
we've set it as an input to the module. However, there's no way in Terraform of
programmatically referring to the currently running version, so we can't set an
automatic default, and you can't call functions when setting the default value
either. So we must maintain the version somewhat by hand.